backend/requirements.txt
Flask==3.0.3
flask_cors==4.0.0
psycopg2-binary==2.9.9
